requirements.txt
flask==3.0.3
flask-cors==4.0.1
garminconnect==0.2.19
python-dotenv==1.0.1
gunicorn==21.2.0
